split pipe to stream and aead parts
